fix: resolved std::chrono & std::thread issues (#84)
- removed nearly all std::chrono usage from the application (still exists in a couple of places to use lambda std::thread instances where lambdas are not supported in QThread on older Qt version). The std::chrono library is cumbersome, and although you gain a known unit for a value, it can be incredibly convoluted to work for. In pingnoo now, if a time duration is shared as an integer, it's milliseconds. As a double, it's seconds, unless otherwise specified.

- fixed issues with threads mixing std::thread and QThread, this caused problems with terminating threads.

- fixed #77, changes were made recently to add singleShot mechanism to ping engines, in singleShot mode an engine can be instantiated and used in a blocking manner without needing to use worker threads for transmission and reception.  ICMPAPIPingEngine had support for singleShot mode, but incorrectly assumed that the threads were created, this caused a crash.

+++ src/components/RouteAnalyser/RouteAnalyserWidget.h
@@ -84,11 +84,13 @@ namespace Nedrysoft { namespace RouteAnalyser {
              * @param[in]   pingEngineFactory the ping engine factory to use.
              * @param[in]   parent the parent widget.
              */
-            explicit RouteAnalyserWidget(QString targetHost,
-                                         Nedrysoft::Core::IPVersion ipVersion,
-                                         double interval,
-                                         Nedrysoft::RouteAnalyser::IPingEngineFactory *pingEngineFactory,
-                                         QWidget *parent = nullptr );
+            explicit RouteAnalyserWidget(
+                QString targetHost,
+                Nedrysoft::Core::IPVersion ipVersion,
+                int interval,
+                Nedrysoft::RouteAnalyser::IPingEngineFactory *pingEngineFactory,
+                QWidget *parent = nullptr
+            );
 
             /**
              * @brief       Destroys the RouteAnalyserWidget.
@@ -219,7 +221,7 @@ namespace Nedrysoft { namespace RouteAnalyser {
             QSplitter *m_splitter;
             PlotScrollArea *m_scrollArea;
             Nedrysoft::RouteAnalyser::IPingEngineFactory *m_pingEngineFactory;
-            double m_interval;
+            int m_interval;
             QList<Nedrysoft::RouteAnalyser::GraphLatencyLayer *> m_backgroundLayers;
             Nedrysoft::RouteAnalyser::RouteTableItemDelegate *m_routeGraphDelegate;
             ScaleMode m_graphScaleMode;
